feat(MerchantContract): seperate update & create forms
